ft: add the util methods for parsing movie metadata struct

+++ src/models.rs
@@ -24,9 +24,9 @@ impl MediaEncodingFormat {
 pub struct MovieMetadata {
     pub(crate) original_file_name: String,
     pub(crate) media_name: String,
-    pub(crate) release_year: Option<u8>,
+    pub(crate) release_year: Option<u16>,
     pub(crate) encoding_format: Option<MediaEncodingFormat>,
-    pub(crate) resolution: Option<u8>,
+    pub(crate) resolution: Option<u16>,
     pub(crate) additional_data: Vec<String>,
     pub(crate) file_extension: String,
 }
